Ported limits to use the constexpr defined pins.
Removed pin inversion defines/settings as this can now be defined directly at
the configuration.

+++ driver/gpio.h
@@ -175,7 +175,7 @@ struct Bus : Detail::Bus {
      * @param port Port number
      * @param pins Initializer list of constexpr pin instances.
      */
-    constexpr Bus(uint32_t port, std::array<Pin, N> pins)
+    constexpr Bus(uint8_t port, std::array<Pin, N> pins)
         : Detail::Bus{port, 0, 0}
         , pins(pins)
     {
